Enum lodepng::Image

pub enum Image {
    RawData(Bitmap<u8>),
    Grey(Bitmap<Grey<u8>>),
    Grey16(Bitmap<Grey<u16>>),
    GreyAlpha(Bitmap<GreyAlpha<u8>>),
    GreyAlpha16(Bitmap<GreyAlpha<u16>>),
    RGBA(Bitmap<RGBA>),
    RGB(Bitmap<RGB<u8>>),
    RGBA16(Bitmap<RGBA<u16>>),
    RGB16(Bitmap<RGB<u16>>),
}

Bitmap types.

Images with >=8bpp are stored with pixel per vec element. Images with <8bpp are represented as a bunch of bytes, with multiple pixels per byte.

Variants
RawData(Bitmap<u8>)

Bytes of the image. See bpp how many pixels per element there are

Grey(Bitmap<Grey<u8>>)Grey16(Bitmap<Grey<u16>>)GreyAlpha(Bitmap<GreyAlpha<u8>>)GreyAlpha16(Bitmap<GreyAlpha<u16>>)RGBA(Bitmap<RGBA>)RGB(Bitmap<RGB<u8>>)RGBA16(Bitmap<RGBA<u16>>)RGB16(Bitmap<RGB<u16>>)