GitHub project page

Struct lodepng::ffi::State

pub struct State {
    pub decoder: DecoderSettings,
    pub encoder: EncoderSettings,
    pub info_raw: ColorMode,
    pub info_png: Info,
    // some fields omitted
}

The settings, state and information for extended encoding and decoding

Fields

decoder

encoder

info_raw

specifies the format in which you would like to get the raw pixel buffer

info_png

info of the PNG image obtained after decoding