Struct lodepng::Bitmap

pub struct Bitmap<PixelType: Copy> {
    pub buffer: Vec<PixelType>,
    pub width: usize,
    pub height: usize,
}

Low-level representation of an image

Fields
buffer: Vec<PixelType>

Raw bitmap memory. Layout depends on color mode and bitdepth used to create it.

For RGB/RGBA images one element is one pixel.

For <8bpp images pixels are packed, so raw bytes are exposed and you need to do bit-twiddling youself

width: usize

Width in pixels

height: usize

Height in pixels