GitHub project page

Struct lodepng::Info

pub struct Info {
    // some fields omitted
}

Methods

impl Info

fn new() -> Info

fn clear_text(&mut self)

use this to clear the texts again after you filled them in

fn add_text(&mut self, key: *const c_char, str: *const c_char) -> Error

push back both texts at once

fn clear_itext(&mut self)

use this to clear the itexts again after you filled them in

fn add_itext(&mut self, key: *const c_char, langtag: *const c_char, transkey: *const c_char, str: *const c_char) -> Error

push back the 4 texts of 1 chunk at once