GitHub project page

Crate lodepng

Reexports

pub use ffi::ColorType;

pub use ffi::CompressSettings;

pub use ffi::Time;

pub use ffi::DecoderSettings;

pub use ffi::FilterStrategy;

pub use ffi::AutoConvert;

pub use ffi::EncoderSettings;

pub use ffi::Error;

Modules

ffi

Structs

Bitmap
Low-level representation of an image

Chunk

ColorMode

Grey
Opaque greyscale pixel (acces with px.0)

GreyAlpha
Greyscale pixel with alpha (px.1 is alpha)

Info

RGB

RGBA

State

Enums

Image
Bitmap types. Also contains valid values for <PixelType>

Functions

decode24
Same as decode_memory, but always decodes to 24-bit RGB raw image

decode24_file
Same as decode_file, but always decodes to 24-bit RGB raw image

decode32
Same as decode_memory, but always decodes to 32-bit RGBA raw image

decode32_file
Same as decode_file, but always decodes to 32-bit RGBA raw image

decode_file
Load PNG from disk, from file with given name. Same as the other decode functions, but instead takes a file path as input.

decode_memory
Converts PNG data in memory to raw pixel data.

encode24
Same as encode_memory, but always encodes from 24-bit RGB raw image

encode24_file
Same as encode_file, but always encodes from 24-bit RGB raw image

encode32
Same as encode_memory, but always encodes from 32-bit RGBA raw image

encode32_file
Same as encode_file, but always encodes from 32-bit RGBA raw image

encode_file
Converts raw pixel data into a PNG file on disk. Same as the other encode functions, but instead takes a file path as output.

encode_memory
Converts raw pixel data into a PNG image in memory. The colortype and bitdepth of the output PNG image cannot be chosen, they are automatically determined by the colortype, bitdepth and content of the input pixel data.