GitHub project page

Module lodepng::ffi

Reexports

pub use cvec::CVec;

Structs

ColorMode
Color mode of an image. Contains all information required to decode the pixel bits to RGBA colors. This information is the same as used in the PNG file format, and is used both for PNG and raw image data in LodePNG.

CompressSettings
Settings for zlib compression. Tweaking these settings tweaks the balance between speed and compression ratio.

DecoderSettings
Settings for the decoder. This contains settings for the PNG and the Zlib decoder, but not the Info settings from the Info structs.

EncoderSettings

Error

Info
Information about the PNG image, except pixels, width and height

State
The settings, state and information for extended encoding and decoding

Time
The information of a Time chunk in PNG

Enums

AutoConvert
automatically use color type with less bits per pixel if losslessly possible. Default: LAC_AUTO

ColorType

FilterStrategy
automatically use color type with less bits per pixel if losslessly possible. Default: AUTO

Functions

lodepng_add_itext

lodepng_add_text

lodepng_auto_choose_color

lodepng_can_have_alpha

lodepng_chunk_ancillary

lodepng_chunk_append

lodepng_chunk_check_crc

lodepng_chunk_create

lodepng_chunk_data

lodepng_chunk_generate_crc

lodepng_chunk_length

lodepng_chunk_next

lodepng_chunk_private

lodepng_chunk_safetocopy

lodepng_chunk_type

lodepng_chunk_type_equals

lodepng_clear_itext

lodepng_clear_text

lodepng_color_mode_cleanup

lodepng_color_mode_copy

lodepng_color_mode_init

lodepng_compress_settings_init

lodepng_convert

lodepng_crc32

lodepng_decode

lodepng_decode_memory

lodepng_decoder_settings_init

lodepng_deflate

lodepng_encode

lodepng_encode_memory

lodepng_encoder_settings_init

lodepng_get_bpp

lodepng_get_channels

lodepng_get_raw_size

lodepng_has_palette_alpha

lodepng_info_cleanup

lodepng_info_copy

lodepng_info_init

lodepng_inspect

lodepng_is_alpha_type

lodepng_is_greyscale_type

lodepng_is_palette_type

lodepng_palette_add

lodepng_palette_clear

lodepng_state_cleanup

lodepng_state_copy

lodepng_state_init

lodepng_zlib_compress