Module lodepng::ffi

Structs

ColorMode
Color mode of an image. Contains all information required to decode the pixel bits to RGBA colors. This information is the same as used in the PNG file format, and is used both for PNG and raw image data in LodePNG.

ColorProfile
Gives characteristics about the colors of the image, which helps decide which color model to use for encoding. Used internally by default if "auto_convert" is enabled. Public because it's useful for custom algorithms.

CompressSettings
Settings for zlib compression. Tweaking these settings tweaks the balance between speed and compression ratio.

DecoderSettings
Settings for the decoder. This contains settings for the PNG and the Zlib decoder, but not the Info settings from the Info structs.

DecompressSettings

EncoderSettings

Error

Info
Information about the PNG image, except pixels, width and height

State
The settings, state and information for extended encoding and decoding

Time
The information of a Time chunk in PNG

Enums

ColorType
Type for decode, encode, etc. Same as standard PNG color types.

FilterStrategy
automatically use color type with less bits per pixel if losslessly possible. Default: AUTO

Functions

lodepng_add_itext⚠

lodepng_add_text⚠

lodepng_auto_choose_color⚠

lodepng_buffer_file⚠

lodepng_can_have_alpha⚠

lodepng_chunk_ancillary⚠

lodepng_chunk_append⚠

lodepng_chunk_check_crc⚠

lodepng_chunk_create⚠

lodepng_chunk_data⚠

lodepng_chunk_data_const⚠

lodepng_chunk_generate_crc⚠

lodepng_chunk_length⚠

lodepng_chunk_next⚠

lodepng_chunk_next_const⚠

lodepng_chunk_private⚠

lodepng_chunk_safetocopy⚠

lodepng_chunk_type⚠

lodepng_chunk_type_equals⚠

lodepng_clear_itext⚠

lodepng_clear_text⚠

lodepng_color_mode_cleanup⚠

lodepng_color_mode_copy⚠

lodepng_color_mode_equal⚠

lodepng_color_mode_init⚠

lodepng_color_profile_init⚠

lodepng_compress_settings_init⚠

lodepng_convert⚠

lodepng_crc32⚠

lodepng_decode⚠

lodepng_decode24⚠

lodepng_decode32⚠

lodepng_decode24_file⚠

lodepng_decode32_file⚠

lodepng_decode_file⚠

lodepng_decode_memory⚠

lodepng_decoder_settings_init⚠

lodepng_decompress_settings_init⚠

lodepng_deflate⚠

lodepng_encode⚠

lodepng_encode24⚠

lodepng_encode32⚠

lodepng_encode24_file⚠

lodepng_encode32_file⚠

lodepng_encode_file⚠

lodepng_encode_memory⚠

lodepng_encoder_settings_init⚠

lodepng_error_text⚠

lodepng_filesize⚠

lodepng_free⚠

lodepng_get_bpp⚠

lodepng_get_bpp_lct⚠

lodepng_get_channels⚠

lodepng_get_color_profile⚠

lodepng_get_raw_size⚠

lodepng_get_raw_size_lct⚠

lodepng_has_palette_alpha⚠

lodepng_huffman_code_lengths⚠

lodepng_inflate⚠

lodepng_info_cleanup⚠

lodepng_info_copy⚠

lodepng_info_init⚠

lodepng_info_swap⚠

lodepng_inspect⚠

lodepng_is_alpha_type⚠

lodepng_is_greyscale_type⚠

lodepng_is_palette_type⚠

lodepng_load_file⚠

lodepng_malloc⚠

lodepng_palette_add⚠

lodepng_palette_clear⚠

lodepng_realloc⚠

lodepng_save_file⚠

lodepng_state_cleanup⚠

lodepng_state_copy⚠

lodepng_state_init⚠

lodepng_zlib_compress⚠

lodepng_zlib_decompress⚠

zlib_compress⚠

zlib_decompress⚠

Type Definitions

custom_compress_callback

custom_decompress_callback