GitHub project page

Enum lodepng::ffi::AutoConvert

pub enum AutoConvert {
    LAC_NO,
    LAC_ALPHA,
    LAC_AUTO,
    LAC_AUTO_NO_NIBBLES,
    LAC_AUTO_NO_PALETTE,
    LAC_AUTO_NO_NIBBLES_NO_PALETTE,
}

automatically use color type with less bits per pixel if losslessly possible. Default: LAC_AUTO

Variants
LAC_NO

use color type user requested

LAC_ALPHA

use color type user requested, but if only opaque pixels and RGBA or grey+alpha, use RGB or grey

LAC_AUTO

use PNG color type that can losslessly represent the uncompressed image the smallest possible

LAC_AUTO_NO_NIBBLES

like AUTO, but do not choose 1, 2 or 4 bit per pixel types. sometimes a PNG image compresses worse if less than 8 bits per pixels.

LAC_AUTO_NO_PALETTE

like AUTO, but never choose palette color type. For small images, encoding the palette may take more bytes than what is gained. Note that AUTO also already prevents encoding the palette for extremely small images, but that may not be sufficient because due to the compression it cannot predict when to switch.

LAC_AUTO_NO_NIBBLES_NO_PALETTE