GitHub project page

Enum lodepng::ffi::FilterStrategyStable

pub enum FilterStrategy {
    LFS_ZERO,
    LFS_MINSUM,
    LFS_ENTROPY,
    LFS_BRUTE_FORCE,
    LFS_PREDEFINED,
}

automatically use color type with less bits per pixel if losslessly possible. Default: AUTO

Variants

LFS_ZERO

every filter at zero

LFS_MINSUM

Use filter that gives minumum sum, as described in the official PNG filter heuristic.

LFS_ENTROPY

Use the filter type that gives smallest Shannon entropy for this scanline. Depending on the image, this is better or worse than minsum.

LFS_BRUTE_FORCE

Brute-force-search PNG filters by compressing each filter for each scanline. Experimental, very slow, and only rarely gives better compression than MINSUM.

LFS_PREDEFINED

use predefined_filters buffer: you specify the filter type for each scanline