Struct lodepng::ffi::EncoderSettings

#[repr(C)]

pub struct EncoderSettings {
    pub zlibsettings: CompressSettings,
    pub auto_convert: c_uint,
    pub filter_palette_zero: c_uint,
    pub filter_strategy: FilterStrategy,
    pub force_palette: c_uint,
    pub add_id: c_uint,
    pub text_compression: c_uint,
    // some fields omitted
}

Fields
zlibsettings: CompressSettings

settings for the zlib encoder, such as window size, ...

auto_convert: c_uint

how to automatically choose output PNG color type, if at all

filter_palette_zero: c_uint

If true, follows the official PNG heuristic: if the PNG uses a palette or lower than 8 bit depth, set all filters to zero. Otherwise use the filter_strategy. Note that to completely follow the official PNG heuristic, filter_palette_zero must be true and filter_strategy must be FilterStrategy::MINSUM

filter_strategy: FilterStrategy

Which filter strategy to use when not using zeroes due to filter_palette_zero. Set filter_palette_zero to 0 to ensure always using your chosen strategy. Default: FilterStrategy::MINSUM

force_palette: c_uint

force creating a PLTE chunk if colortype is 2 or 6 (= a suggested palette). If colortype is 3, PLTE is always created

add_id: c_uint

add LodePNG identifier and version as a text chunk, for debugging

text_compression: c_uint

encode text chunks as zTXt chunks instead of tEXt chunks, and use compression in iTXt chunks

Methods

impl EncoderSettings

pub fn new() -> Self