GitHub project page

Struct lodepng::ffi::EncoderSettings

pub struct EncoderSettings {
    pub zlibsettings: CompressSettings,
    pub auto_convert: AutoConvert,
    pub filter_palette_zero: c_uint,
    pub filter_strategy: FilterStrategy,
    pub force_palette: c_uint,
    // some fields omitted
}

Fields

zlibsettings

settings for the zlib encoder, such as window size, ...

auto_convert

how to automatically choose output PNG color type, if at all

filter_palette_zero

If true, follows the official PNG heuristic: if the PNG uses a palette or lower than 8 bit depth, set all filters to zero. Otherwise use the filter_strategy. Note that to completely follow the official PNG heuristic, filter_palette_zero must be true and filter_strategy must be LFS_MINSUM

filter_strategy

Which filter strategy to use when not using zeroes due to filter_palette_zero. Set filter_palette_zero to 0 to ensure always using your chosen strategy. Default: LFS_MINSUM

force_palette

force creating a PLTE chunk if colortype is 2 or 6 (= a suggested palette). If colortype is 3, PLTE is always created

Methods

impl EncoderSettings

fn new() -> EncoderSettings

Creates encoder settings initialized to defaults