GitHub project page

Struct lodepng::ffi::Info

pub struct Info {
    pub compression_method: c_uint,
    pub filter_method: c_uint,
    pub interlace_method: c_uint,
    pub color: ColorMode,
    pub background_defined: c_uint,
    pub background_r: c_uint,
    pub background_g: c_uint,
    pub background_b: c_uint,
    pub time_defined: c_uint,
    pub time: Time,
    pub phys_defined: c_uint,
    pub phys_x: c_uint,
    pub phys_y: c_uint,
    pub phys_unit: c_uint,
    // some fields omitted
}

Information about the PNG image, except pixels, width and height

Fields

compression_method

compression method of the original file. Always 0.

filter_method

filter method of the original file

interlace_method

interlace method of the original file

color

color type and bits, palette and transparency of the PNG file

background_defined

suggested background color chunk (bKGD) This color uses the same color mode as the PNG (except alpha channel), which can be 1-bit to 16-bit.

For greyscale PNGs, r, g and b will all 3 be set to the same. When encoding the encoder writes the red one. For palette PNGs: When decoding, the RGB value will be stored, not a palette index. But when encoding, specify the index of the palette in background_r, the other two are then ignored.

The decoder does not use this background color to edit the color of pixels.

background_r

red component of suggested background color

background_g

green component of suggested background color

background_b

blue component of suggested background color

time_defined

time chunk (tIME)

time

set to 1 to make the encoder generate a tIME chunk

phys_defined

if 0, there is no pHYs chunk and the values below are undefined, if 1 else there is one

phys_x

pixels per unit in x direction

phys_y

pixels per unit in y direction

phys_unit

may be 0 (unknown unit) or 1 (metre)

Methods

impl Info

fn new() -> Info

fn clear_text(&mut self)

use this to clear the texts again after you filled them in

fn add_text(&mut self, key: *const c_char, str: *const c_char) -> Error

push back both texts at once

fn clear_itext(&mut self)

use this to clear the itexts again after you filled them in

fn add_itext(&mut self, key: *const c_char, langtag: *const c_char, transkey: *const c_char, str: *const c_char) -> Error

push back the 4 texts of 1 chunk at once