Struct lodepng::ffi::State

#[repr(C)]

pub struct State {
    pub decoder: DecoderSettings,
    pub encoder: EncoderSettings,
    pub info_raw: ColorMode,
    pub info_png: Info,
    pub error: Error,
}

The settings, state and information for extended encoding and decoding

Fields
decoder: DecoderSettings encoder: EncoderSettings info_raw: ColorMode

specifies the format in which you would like to get the raw pixel buffer

info_png: Info

info of the PNG image obtained after decoding

error: Error
Methods

impl State

pub fn new() -> Self

pub fn set_auto_convert(&mut self, mode: bool)

pub fn set_filter_strategy(
&mut self,
mode: FilterStrategy,
palette_filter_zero: bool
)

pub unsafe fn set_custom_zlib(
&mut self,
callback: custom_compress_callback,
context: *const c_void
)

pub unsafe fn set_custom_deflate(
&mut self,
callback: custom_compress_callback,
context: *const c_void
)

pub fn info_raw(&self) -> &ColorMode

pub fn info_raw_mut(&mut self) -> &mut ColorMode

pub fn info_png(&self) -> &Info

pub fn info_png_mut(&mut self) -> &mut Info

pub fn color_convert(&mut self, true_or_false: bool)

whether to convert the PNG to the color type you want. Default: yes

pub fn read_text_chunks(&mut self, true_or_false: bool)

if false but remember_unknown_chunks is true, they're stored in the unknown chunks.

pub fn remember_unknown_chunks(&mut self, true_or_false: bool)

store all bytes from unknown chunks in the LodePNGInfo (off by default, useful for a png editor)

pub fn get_icc(&self) -> Result<Vec<u8>, Error>

Decompress ICC profile from iCCP chunk

pub fn decode<Bytes: AsRef<[u8]>>(
&mut self,
input: Bytes
) -> Result<Image, Error>

Load PNG from buffer using State's settings

 state.info_raw_mut().colortype = ColorType::RGBA;
 match state.decode(&slice) {
     Ok(Image::RGBA(with_alpha)) => do_stuff(with_alpha),
     _ => panic!("¯\\_(ツ)_/¯")
 }

pub fn decode_file<P: AsRef<Path>>(
&mut self,
filepath: P
) -> Result<Image, Error>

pub fn inspect(&mut self, input: &[u8]) -> Result<(usize, usize), Error>

Returns (width, height)

pub fn encode<PixelType: Copy>(
&mut self,
image: &[PixelType],
w: usize,
h: usize
) -> Result<Vec<u8>, Error>

pub fn encode_file<PixelType: Copy, P: AsRef<Path>>(
&mut self,
filepath: P,
image: &[PixelType],
w: usize,
h: usize
) -> Result<(), Error>